+++ src/sdk/PnP.Core/Model/AzureActiveDirectory/Internal/Group.cs
@@ -1,0 +1,8 @@
+﻿namespace PnP.Core.Model.AzureActiveDirectory
+{
+    [GraphType(Get = "groups/{GraphId}")]
+    internal partial class Group
+    {
+
+    }
+}