+++ src/sdk/PnP.Core/Model/Base/BaseDataModel.cs
@@ -667,13 +667,13 @@ namespace PnP.Core.Model
                     throw new ClientException(ErrorType.ModelMetadataIncorrect, $"Specify the GraphGet field of the ClassMapping property.");
                 }
 
-                // TODO: Consider replacing this code with a call to the TokenHandler
-                baseApiCall = entity.GraphGet.Replace("{hostname}", PnPContext.Uri.DnsSafeHost).Replace("{serverrelativepath}", PnPContext.Uri.PathAndQuery);
+                // Ensure tokens in the base url are replaced
+                baseApiCall = TokenHandler.ResolveTokensAsync(this, entity.GraphGet, PnPContext).GetAwaiter().GetResult();
             }
             else
             {
-                // TODO: Consider replacing this code with a call to the TokenHandler
-                baseApiCall = apiOverride.Request.Replace("{hostname}", PnPContext.Uri.DnsSafeHost).Replace("{serverrelativepath}", PnPContext.Uri.PathAndQuery);
+                // Ensure tokens in the base url are replaced
+                baseApiCall = TokenHandler.ResolveTokensAsync(this, apiOverride.Request, PnPContext).GetAwaiter().GetResult();
             }
 
             // Parse tokens in the base api call